# packages/api-client/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "hvnt33 API",
    "version": "0.2.0",
    "description": "Investigations, evidence, intake, observed search results, visits and web archive. Local mode trusts loopback requests; token mode requires a bearer API token and scopes every request to its workspace."
  },
  "servers": [
    {
      "url": "http://localhost:4310",
      "description": "Local hvnt33 server"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/records/{id}": {
      "patch": {
        "summary": "Edit a record",
        "tags": [
          "Records"
        ],
        "operationId": "patchApiRecordsId",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 100,
              "description": "Identifier"
            },
            "description": "Identifier"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string",
                    "maxLength": 1000
                  },
                  "notes": {
                    "type": "string",
                    "maxLength": 200000
                  },
                  "sourceLabel": {
                    "type": "string",
                    "maxLength": 1000
                  },
                  "tags": {
                    "type": "string",
                    "maxLength": 4000
                  },
                  "eventDate": {
                    "type": "string",
                    "maxLength": 20
                  },
                  "sourceUrl": {
                    "type": "string",
                    "maxLength": 4000
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "Unverified",
                      "Corroborated",
                      "Verified",
                      "Disputed"
                    ]
                  },
                  "public": {
                    "type": "boolean"
                  },
                  "reviewed": {
                    "description": "true: a person has reviewed this (recorded with who and when); false: clear it",
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "workspaceId": {
                      "type": "string"
                    },
                    "investigationId": {
                      "type": "string"
                    },
                    "title": {
                      "type": "string"
                    },
                    "kind": {
                      "type": "string",
                      "enum": [
                        "Person",
                        "Organization",
                        "Place",
                        "Event",
                        "Claim",
                        "Document",
                        "Image",
                        "Video",
                        "Link",
                        "Note"
                      ]
                    },
                    "status": {
                      "type": "string",
                      "enum": [
                        "Unverified",
                        "Corroborated",
                        "Verified",
                        "Disputed"
                      ]
                    },
                    "notes": {
                      "type": "string"
                    },
                    "sourceUrl": {
                      "type": "string"
                    },
                    "sourceLabel": {
                      "type": "string"
                    },
                    "eventDate": {
                      "type": "string"
                    },
                    "tags": {
                      "type": "string"
                    },
                    "public": {
                      "type": "boolean"
                    },
                    "fileKey": {
                      "type": "string"
                    },
                    "filename": {
                      "type": "string"
                    },
                    "mime": {
                      "type": "string"
                    },
                    "size": {
                      "type": "number"
                    },
                    "sha256": {
                      "type": "string"
                    },
                    "sourceQuote": {
                      "type": "string"
                    },
                    "sourceId": {
                      "type": "string"
                    },
                    "intakeId": {
                      "type": "string"
                    },
                    "filedBy": {
                      "type": "string"
                    },
                    "contentOrigin": {
                      "type": "string"
                    },
                    "createdAt": {
                      "type": "string"
                    },
                    "updatedAt": {
                      "type": "string"
                    },
                    "version": {
                      "type": "number"
                    },
                    "deletedAt": {
                      "type": "string"
                    },
                    "reviewedAt": {
                      "description": "When a person last reviewed this; empty or absent: never",
                      "type": "string"
                    },
                    "reviewedBy": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "id",
                    "workspaceId",
                    "investigationId",
                    "title",
                    "kind",
                    "status",
                    "notes",
                    "sourceUrl",
                    "sourceLabel",
                    "eventDate",
                    "tags",
                    "public",
                    "createdAt",
                    "updatedAt",
                    "version",
                    "deletedAt"
                  ],
                  "additionalProperties": {},
                  "description": "A person, organization, place, event, claim, document, image, video, link or note"
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete a record (kept as a tombstone for sync; its connections are deleted too)",
        "tags": [
          "Records"
        ],
        "operationId": "deleteApiRecordsId",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 100,
              "description": "Identifier"
            },
            "description": "Identifier"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "deleted": {
                      "type": "string"
                    },
                    "connections": {
                      "type": "number"
                    }
                  },
                  "required": [
                    "deleted",
                    "connections"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "API token (token mode). Local mode trusts loopback requests and needs no token."
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "details": {
            "type": "object"
          }
        }
      }
    }
  }
}